requirements.txt
openai>=1.40.0
google-generativeai>=0.8.0
PyMuPDF>=1.24.0
pydub>=0.25.1
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
